[helpers] modularized helper functions into rules/helpers
