--- ui/client/rules/find-king-square.js
@@ -1,9 +1,0 @@
-module.exports = (king, matrix) => {
-  for (let i = 0; i < matrix.length; i++) {
-    for (let j = 0; j < matrix[i].length; j++) {
-      if (matrix[i][j] === king) {
-        return [i,j];
-      }
-    }
-  }
-}
