--- ui/client/rules/is-white.js
@@ -1,4 +1,0 @@
-module.exports = (piece) => {
-  return piece === null ? null :
-    piece === piece.toUpperCase() ? true : false;
-}
